networking: remove implementation specific iface type
The caller code is not implemented to handle different types than char.
And the implementations does not provide any interface to convert from
char to the implementation specific type.

So the speficic type is actually not supported.

This removes the specific type as a generic support is currently not
possible to have a different type.

Signed-off-by: Gaëtan Harter <hartergaetan@gmail.com>